Make a plugin framework and move over trigger.
