+++ prow/cmd/hook/main.go
@@ -21,6 +21,7 @@ import (
 	"flag"
 	"io/ioutil"
 	"net/http"
+	"os"
 	"os/signal"
 	"strconv"
 	"syscall"
@@ -30,19 +31,17 @@ import (
 	"k8s.io/test-infra/prow/github"
 	"k8s.io/test-infra/prow/jobs"
 	"k8s.io/test-infra/prow/kube"
+	"k8s.io/test-infra/prow/plugins"
+
+	_ "k8s.io/test-infra/prow/plugins/lgtm"
+	_ "k8s.io/test-infra/prow/plugins/trigger"
 )
 
 var (
-	port      = flag.Int("port", 8888, "Port to listen on.")
-	namespace = flag.String("namespace", "default", "Namespace for all CI objects.")
-	dryRun    = flag.Bool("dry-run", true, "Whether or not to avoid mutating calls to GitHub.")
-	org       = flag.String("org", "kubernetes", "GitHub org to trust.")
+	port = flag.Int("port", 8888, "Port to listen on.")
 
-	jobConfig    = flag.String("job-config", "/etc/jobs/jobs", "Path to job config file.")
-	pluginConfig = flag.String("plugin-config", "/etc/plugins/plugins", "Path to plugin config file.")
-
-	lineImage = flag.String("line-image", "", "Image to use for testing PRs.")
-
+	jobConfig         = flag.String("job-config", "/etc/jobs/jobs", "Path to job config file.")
+	pluginConfig      = flag.String("plugin-config", "/etc/plugins/plugins", "Path to plugin config file.")
 	webhookSecretFile = flag.String("hmac-secret-file", "/etc/webhook/hmac", "Path to the file containing the GitHub HMAC secret.")
 	githubTokenFile   = flag.String("github-token-file", "/etc/github/oauth", "Path to the file containing the GitHub OAuth secret.")
 )
@@ -51,77 +50,68 @@ func main() {
 	flag.Parse()
 	logrus.SetFormatter(&logrus.JSONFormatter{})
 
-	webhookSecretRaw, err := ioutil.ReadFile(*webhookSecretFile)
-	if err != nil {
-		logrus.Fatalf("Could not read webhook secret file: %s", err)
-	}
-	webhookSecret := bytes.TrimSpace(webhookSecretRaw)
-
-	oauthSecretRaw, err := ioutil.ReadFile(*githubTokenFile)
-	if err != nil {
-		logrus.Fatalf("Could not read oauth secret file: %s", err)
-	}
-	oauthSecret := string(bytes.TrimSpace(oauthSecretRaw))
-
-	var githubClient *github.Client
-	if *dryRun {
-		githubClient = github.NewDryRunClient(oauthSecret)
-	} else {
-		githubClient = github.NewClient(oauthSecret)
-	}
-
-	kubeClient, err := kube.NewClientInCluster(*namespace)
-	if err != nil {
-		logrus.Fatalf("Error getting client: %s", err)
-	}
-
 	// Ignore SIGTERM so that we don't drop hooks when the pod is removed.
 	// We'll get SIGTERM first and then SIGKILL after our graceful termination
 	// deadline.
 	signal.Ignore(syscall.SIGTERM)
 
+	webhookSecretRaw, err := ioutil.ReadFile(*webhookSecretFile)
+	if err != nil {
+		logrus.WithError(err).Fatal("Could not read webhook secret file.")
+	}
+	webhookSecret := bytes.TrimSpace(webhookSecretRaw)
+
+	oauthSecretRaw, err := ioutil.ReadFile(*githubTokenFile)
+	if err != nil {
+		logrus.WithError(err).Fatal("Could not read oauth secret file.")
+	}
+	oauthSecret := string(bytes.TrimSpace(oauthSecretRaw))
+
+	dry, err := strconv.ParseBool(os.Getenv("DRY_RUN"))
+	if err != nil {
+		logrus.WithError(err).Fatal("Failed to parse DRY_RUN environment variable.")
+	}
+
+	var githubClient *github.Client
+	if dry {
+		githubClient = github.NewDryRunClient(oauthSecret)
+	} else {
+		githubClient = github.NewClient(oauthSecret)
+	}
+
+	kubeClient, err := kube.NewClientInCluster("default")
+	if err != nil {
+		logrus.WithError(err).Fatal("Error getting kube client.")
+	}
+
+	jobAgent := &jobs.JobAgent{}
+	if err := jobAgent.Start(*jobConfig); err != nil {
+		logrus.WithError(err).Fatal("Error starting job agent.")
+	}
+
+	pluginAgent := &plugins.PluginAgent{
+		GitHubClient: githubClient,
+		KubeClient:   kubeClient,
+		JobAgent:     jobAgent,
+	}
+	if err := pluginAgent.Start(*pluginConfig); err != nil {
+		logrus.WithError(err).Fatal("Error starting plugins.")
+	}
+
 	prc := make(chan github.PullRequestEvent)
 	icc := make(chan github.IssueCommentEvent)
-	brc := make(chan KubeRequest)
-	drc := make(chan KubeRequest)
 	server := &Server{
 		HMACSecret:         webhookSecret,
 		PullRequestEvents:  prc,
 		IssueCommentEvents: icc,
 	}
 
-	jobAgent := &jobs.JobAgent{}
-	jobAgent.Start(*jobConfig)
-
-	pluginAgent := &PluginAgent{}
-	pluginAgent.Start(*pluginConfig)
-
-	githubAgent := &GitHubAgent{
-		DryRun:       *dryRun,
-		Org:          *org,
-		GitHubClient: githubClient,
-
-		JenkinsJobs: jobAgent,
-		Plugins:     pluginAgent,
-
+	events := &EventAgent{
+		Plugins:            pluginAgent,
 		PullRequestEvents:  prc,
 		IssueCommentEvents: icc,
-
-		BuildRequests:  brc,
-		DeleteRequests: drc,
 	}
-	githubAgent.Start()
-
-	kubeAgent := &KubeAgent{
-		DryRun:     *dryRun,
-		LineImage:  *lineImage,
-		KubeClient: kubeClient,
-		Namespace:  *namespace,
-
-		BuildRequests:  brc,
-		DeleteRequests: drc,
-	}
-	kubeAgent.Start()
+	events.Start()
 
 	logrus.Fatal(http.ListenAndServe(":"+strconv.Itoa(*port), server))
 }
