+++ prow/cmd/hook/main_test.go
@@ -1,0 +1,31 @@
+/*
+Copyright 2016 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package main
+
+import (
+	"testing"
+
+	"k8s.io/test-infra/prow/plugins"
+)
+
+// Make sure that our plugins are valid.
+func TestPlugins(t *testing.T) {
+	pa := &plugins.PluginAgent{}
+	if err := pa.Load("../../plugins.yaml"); err != nil {
+		t.Fatalf("Could not load plugins: %v. Did you forget an import in cmd/hook?", err)
+	}
+}
